Move (c)format basic implementations to rustpython-common
