+++ common/Cargo.toml
@@ -9,8 +9,10 @@ threading = ["parking_lot"]
 
 [dependencies]
 ascii = "1.0"
+bitflags = "1.3.2"
 cfg-if = "1.0"
 hexf-parse = "0.2.1"
+itertools = "0.10.3"
 lexical-parse-float = { version = "0.8.0", features = ["format"] }
 libc = "0.2.133"
 lock_api = "0.4"
